// src/domain/enums/instagram.enum.ts
export enum InstagramConnectedAccountStatusEnum {
  CONNECTED = "connected",
  EXPIRED = "expired",
  DISCONNECTED = "disconnected",
}

export enum PublicationTypeEnum {
  POST = "post",
  STORY = "story",
}

export enum PublicationDestinationScopeEnum {
  ALL = "all",
  SELECTED = "selected",
}

export enum PublicationStatusEnum {
  PENDING = "pending",
  PROCESSING = "processing",
  COMPLETED = "completed",
  PARTIAL_FAILURE = "partial_failure",
  FAILED = "failed",
}

export enum PublicationTargetStatusEnum {
  PENDING = "pending",
  PROCESSING = "processing",
  SUCCESS = "success",
  FAILED = "failed",
}